+++ Assets/Scripts/TileEntity/Interfaces/TileEntityInstanceInterfaces.cs
@@ -30,7 +30,7 @@ namespace TileEntity {
     /// TileEntityObjects which implement this interface can be their instances climbed by the player
     /// </summary>
     public interface IClimableTileEntity {
-        public int GetSpeed();
+        public float GetSpeed();
     }
 
     /// <summary>
